requirements.txt
# Core Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Voice & Audio
streamlit-mic-recorder>=0.0.4
speech-recognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11

# AI & LLM
google-generativeai>=0.3.2

# Data Processing
pandas>=2.1.0
plotly>=5.17.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
streamlit-chat>=0.1.1

# Logging & Monitoring
loguru>=0.7.0
